Updated all package command profiles to return a vector instead of a statically assigned array. This makes the main function more stable as programs vary in update steps. This change allows for several different methods to be used and safely return the return code.
This is a stepper update and shows lots of errors. However, the changes are stable and do not return errors. The main issue is within the main function regarding how the output is captured.

The final version of this new release allows for reporting at the end of the program based upon the loaded profiles.